added make fre in makefil and changed make re, made wall_width a float

+++ src/render2.c
@@ -6,7 +6,7 @@
 /*   By: lcouturi <lcouturi@student.42quebec>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/23 09:12:02 by lcouturi          #+#    #+#             */
-/*   Updated: 2024/03/17 17:09:47 by fbarrett         ###   ########.fr       */
+/*   Updated: 2024/03/18 07:08:56 by fbarrett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ float	*ft_getscale(t_cube c, float screenx, int *i)
 	static float	r[3];
 	float			x;
 	float			y;
-	int				wall_width;
+	float			wall_width;
 
 	angle = c.rotation - FOV / 2.0 + (float)FOV / c.mlx->width * screenx;
 	if (angle < 0)
@@ -72,17 +72,18 @@ float	*ft_getscale(t_cube c, float screenx, int *i)
 	teta = -(FOV / 2.0) + (float)FOV / c.mlx->width * screenx;
 	opposite = cos(fabs(teta) * (M_PI / 180)) * hypothenuse;
 	r[0] = (float)SIZE * c.mlx->width / opposite;
+	//r[0] = (float)SIZE * c.mlx->width / hypothenuse;
 	wall_width = c.map->walls[(int)r[1]][i[(int)r[1]]]->width;
 	if (c.map->map[(int)y / SIZE][(int)x / SIZE] == 'D')
 		wall_width = c.map->walls[4][i[4]]->width;
 	if (!r[1])
-		r[2] = (float)wall_width / SIZE * fmod(x, SIZE);
+		r[2] = wall_width / SIZE * fmod(x, SIZE);
 	else if (r[1] == 1)
-		r[2] = wall_width - (float)wall_width / SIZE * fmod(x, SIZE);
+		r[2] = wall_width - wall_width / SIZE * fmod(x, SIZE);
 	else if (r[1] == 2)
-		r[2] = wall_width - (float)wall_width / SIZE * fmod(y, SIZE);
+		r[2] = wall_width - wall_width / SIZE * fmod(y, SIZE);
 	else
-		r[2] = (float)wall_width / SIZE * fmod(y, SIZE);
+		r[2] = wall_width / SIZE * fmod(y, SIZE);
 	if (c.map->map[(int)y / SIZE][(int)x / SIZE] == 'D')
 		r[1] = 4;
 	return (r);
